Refactoring and Review Changes
