+++ internal/model/holiday.go
@@ -8,12 +8,3 @@ type Holiday struct {
 	HolidayType string    `db:"type" json:"type"`
 	HolidayDate time.Time `db:"date" json:"date"`
 }
-
-type TimeUnit string
-
-const (
-	Epoch       TimeUnit = "Epoch"
-	Month       TimeUnit = "Month"
-	DateOfMonth TimeUnit = "DateOfMonth"
-	Year        TimeUnit = "Year"
-)
